+++ core/CHANGELOG.rst
@@ -1,4 +1,11 @@
-0.2
+0.4.dev
+-------
+
+- change "sendmail" and "process-outoing" commands to not add autocrypt
+  headers if no identity can be determined for a mail.
+
+
+0.3
 ---
 
 - initial "autocrypt" command line client with subcommands.